% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diHeatmap.R
\name{diHeatmap}
\alias{diHeatmap}
\title{diHeatmap}
\usage{
diHeatmap(top = top, bottom = bottom, topCol = c("#FFFADD",
  "#AB221F"), bottomCol = c("#FFFADD", "#3878C1"))
}
\arguments{
\item{top}{the top left heatmap matrix}

\item{bottom}{the bottom right heatmap matrix}

\item{topCol}{the color of top left heatmap}

\item{bottomCol}{the color of bottom right heatmap}
}
\description{
diHeatmap
}
\examples{
data(indel)
diHeatmap(top, bottom)
}
